# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface UserSizes {
  top?: string;
  bottom?: string;
  shoes?: string;
  dress?: string;
}

export type UserProfileGender =
  (typeof UserProfileGender)[keyof typeof UserProfileGender];

export const UserProfileGender = {
  woman: "woman",
  man: "man",
  "non-binary": "non-binary",
  "prefer-not-to-say": "prefer-not-to-say",
} as const;

export type UserProfileSkinTone =
  (typeof UserProfileSkinTone)[keyof typeof UserProfileSkinTone];

export const UserProfileSkinTone = {
  fair: "fair",
  light: "light",
  medium: "medium",
  olive: "olive",
  tan: "tan",
  deep: "deep",
  rich: "rich",
} as const;

export interface UserProfile {
  /** UUID that identifies the device/user */
  profileId: string;
  name?: string;
  gender?: UserProfileGender;
  age?: number;
  skinTone?: UserProfileSkinTone;
  /** City or region for local brand discovery */
  location?: string;
  sizes?: UserSizes;
  /** Free-form style keywords the user loves */
  stylePreferences?: string[];
  /** Things the user wants to avoid */
  avoidKeywords?: string[];
  likedLookIds?: string[];
  createdAt?: string;
  updatedAt?: string;
}

export type UpsertProfileRequestGender =
  (typeof UpsertProfileRequestGender)[keyof typeof UpsertProfileRequestGender];

export const UpsertProfileRequestGender = {
  woman: "woman",
  man: "man",
  "non-binary": "non-binary",
  "prefer-not-to-say": "prefer-not-to-say",
} as const;

export type UpsertProfileRequestSkinTone =
  (typeof UpsertProfileRequestSkinTone)[keyof typeof UpsertProfileRequestSkinTone];

export const UpsertProfileRequestSkinTone = {
  fair: "fair",
  light: "light",
  medium: "medium",
  olive: "olive",
  tan: "tan",
  deep: "deep",
  rich: "rich",
} as const;

export interface UpsertProfileRequest {
  profileId: string;
  name?: string;
  gender?: UpsertProfileRequestGender;
  age?: number;
  skinTone?: UpsertProfileRequestSkinTone;
  location?: string;
  sizes?: UserSizes;
  stylePreferences?: string[];
  avoidKeywords?: string[];
  likedLookIds?: string[];
}

export interface GenerateOutfitsRequest {
  prompt: string;
  numLooks?: number;
  /** Maximum total spend for a single look in USD */
  maxBudget?: number;
  userSizes?: UserSizes;
  userProfile?: UserProfile;
}

export interface OutfitItem {
  /** Top, Bottom, Dress, Shoes, Bag, Jewelry, Accessories, Makeup, Hair */
  category: string;
  name: string;
  brand: string;
  description: string;
  price: number;
  purchaseUrl: string;
  imageUrl?: string;
  color: string;
}

export interface ItemHotspot {
  itemIndex: number;
  category: string;
  xPct: number;
  yPct: number;
}

export interface OutfitLook {
  id: string;
  title: string;
  vibe: string;
  totalCost: number;
  items: OutfitItem[];
  styleNotes: string;
  modelImageB64?: string;
  hotspots?: ItemHotspot[];
}

export interface GenerateModelImageRequest {
  look: OutfitLook;
  userSizes?: UserSizes;
  userProfile?: UserProfile;
}

export interface GenerateModelImageResponse {
  lookId: string;
  modelImageB64: string;
  hotspots: ItemHotspot[];
}

export interface GenerateOutfitsResponse {
  prompt: string;
  looks: OutfitLook[];
  userSizes?: UserSizes;
  maxBudget?: number;
}

export interface SaveOutfitRequest {
  prompt: string;
  look: OutfitLook;
  userSizes?: UserSizes;
}

export interface SavedOutfit {
  id: number;
  prompt: string;
  look: OutfitLook;
  userSizes?: UserSizes;
  savedAt: string;
}

export interface SavedOutfitsResponse {
  outfits: SavedOutfit[];
}

export type GetProfileParams = {
  profileId: string;
};