# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Fit Finder API
servers:
  - url: /api
tags:
  - name: health
  - name: outfits
  - name: profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      parameters:
        - name: profileId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileRequest"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /outfits/generate:
    post:
      operationId: generateOutfits
      tags: [outfits]
      summary: Generate outfit looks with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOutfitsRequest"
      responses:
        "200":
          description: Generated outfit looks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOutfitsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /outfits/model-image:
    post:
      operationId: generateModelImage
      tags: [outfits]
      summary: Generate a model image for a look
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateModelImageRequest"
      responses:
        "200":
          description: Generated model image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateModelImageResponse"

  /outfits/saved:
    get:
      operationId: getSavedOutfits
      tags: [outfits]
      summary: Get saved outfits
      responses:
        "200":
          description: List of saved outfits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfitsResponse"

  /outfits/saved/{id}:
    get:
      operationId: getSavedOutfit
      tags: [outfits]
      summary: Get a saved outfit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved outfit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteSavedOutfit
      tags: [outfits]
      summary: Delete a saved outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /outfits/save:
    post:
      operationId: saveOutfit
      tags: [outfits]
      summary: Save an outfit look
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveOutfitRequest"
      responses:
        "200":
          description: Outfit saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfit"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    UserSizes:
      type: object
      properties:
        top:
          type: string
        bottom:
          type: string
        shoes:
          type: string
        dress:
          type: string

    UserProfile:
      type: object
      properties:
        profileId:
          type: string
          description: UUID that identifies the device/user
        name:
          type: string
        gender:
          type: string
          enum: [woman, man, non-binary, prefer-not-to-say]
        age:
          type: integer
        skinTone:
          type: string
          enum: [fair, light, medium, olive, tan, deep, rich]
        location:
          type: string
          description: City or region for local brand discovery
        sizes:
          $ref: "#/components/schemas/UserSizes"
        stylePreferences:
          type: array
          items:
            type: string
          description: Free-form style keywords the user loves
        avoidKeywords:
          type: array
          items:
            type: string
          description: Things the user wants to avoid
        likedLookIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [profileId]

    UpsertProfileRequest:
      type: object
      properties:
        profileId:
          type: string
        name:
          type: string
        gender:
          type: string
          enum: [woman, man, non-binary, prefer-not-to-say]
        age:
          type: integer
        skinTone:
          type: string
          enum: [fair, light, medium, olive, tan, deep, rich]
        location:
          type: string
        sizes:
          $ref: "#/components/schemas/UserSizes"
        stylePreferences:
          type: array
          items:
            type: string
        avoidKeywords:
          type: array
          items:
            type: string
        likedLookIds:
          type: array
          items:
            type: string
      required: [profileId]

    GenerateOutfitsRequest:
      type: object
      properties:
        prompt:
          type: string
        numLooks:
          type: integer
          default: 4
        maxBudget:
          type: number
          description: Maximum total spend for a single look in USD
        userSizes:
          $ref: "#/components/schemas/UserSizes"
        userProfile:
          $ref: "#/components/schemas/UserProfile"
      required: [prompt]

    GenerateModelImageRequest:
      type: object
      properties:
        look:
          $ref: "#/components/schemas/OutfitLook"
        userSizes:
          $ref: "#/components/schemas/UserSizes"
        userProfile:
          $ref: "#/components/schemas/UserProfile"
      required: [look]

    GenerateModelImageResponse:
      type: object
      properties:
        lookId:
          type: string
        modelImageB64:
          type: string
        hotspots:
          type: array
          items:
            $ref: "#/components/schemas/ItemHotspot"
      required: [lookId, modelImageB64, hotspots]

    ItemHotspot:
      type: object
      properties:
        itemIndex:
          type: integer
        category:
          type: string
        xPct:
          type: number
        yPct:
          type: number
      required: [itemIndex, category, xPct, yPct]

    OutfitItem:
      type: object
      properties:
        category:
          type: string
          description: "Top, Bottom, Dress, Shoes, Bag, Jewelry, Accessories, Makeup, Hair"
        name:
          type: string
        brand:
          type: string
        description:
          type: string
        price:
          type: number
        purchaseUrl:
          type: string
        imageUrl:
          type: string
        color:
          type: string
      required: [category, name, brand, description, price, purchaseUrl, color]

    OutfitLook:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        vibe:
          type: string
        totalCost:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OutfitItem"
        styleNotes:
          type: string
        modelImageB64:
          type: string
        hotspots:
          type: array
          items:
            $ref: "#/components/schemas/ItemHotspot"
      required: [id, title, vibe, totalCost, items, styleNotes]

    GenerateOutfitsResponse:
      type: object
      properties:
        prompt:
          type: string
        looks:
          type: array
          items:
            $ref: "#/components/schemas/OutfitLook"
        userSizes:
          $ref: "#/components/schemas/UserSizes"
        maxBudget:
          type: number
      required: [prompt, looks]

    SaveOutfitRequest:
      type: object
      properties:
        prompt:
          type: string
        look:
          $ref: "#/components/schemas/OutfitLook"
        userSizes:
          $ref: "#/components/schemas/UserSizes"
      required: [prompt, look]

    SavedOutfit:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        look:
          $ref: "#/components/schemas/OutfitLook"
        userSizes:
          $ref: "#/components/schemas/UserSizes"
        savedAt:
          type: string
          format: date-time
      required: [id, prompt, look, savedAt]

    SavedOutfitsResponse:
      type: object
      properties:
        outfits:
          type: array
          items:
            $ref: "#/components/schemas/SavedOutfit"
      required: [outfits]