# lib/api-zod/src/generated/types/generateModelImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { OutfitLook } from "./outfitLook";
import type { UserProfile } from "./userProfile";
import type { UserSizes } from "./userSizes";

export interface GenerateModelImageRequest {
  look: OutfitLook;
  userSizes?: UserSizes;
  userProfile?: UserProfile;
}