# lib/api-zod/src/generated/types/generateModelImageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { ItemHotspot } from "./itemHotspot";

export interface GenerateModelImageResponse {
  lookId: string;
  modelImageB64: string;
  hotspots: ItemHotspot[];
}