# lib/api-zod/src/generated/types/generateOutfitsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { UserProfile } from "./userProfile";
import type { UserSizes } from "./userSizes";

export interface GenerateOutfitsRequest {
  prompt: string;
  numLooks?: number;
  /** Maximum total spend for a single look in USD */
  maxBudget?: number;
  userSizes?: UserSizes;
  userProfile?: UserProfile;
}