# lib/api-zod/src/generated/types/generateOutfitsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { OutfitLook } from "./outfitLook";
import type { UserSizes } from "./userSizes";

export interface GenerateOutfitsResponse {
  prompt: string;
  looks: OutfitLook[];
  userSizes?: UserSizes;
  maxBudget?: number;
}