# lib/api-zod/src/generated/types/itemHotspot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */

export interface ItemHotspot {
  itemIndex: number;
  category: string;
  xPct: number;
  yPct: number;
}