# lib/api-zod/src/generated/types/outfitItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */

export interface OutfitItem {
  /** Top, Bottom, Dress, Shoes, Bag, Jewelry, Accessories, Makeup, Hair */
  category: string;
  name: string;
  brand: string;
  description: string;
  price: number;
  purchaseUrl: string;
  imageUrl?: string;
  color: string;
}