# lib/api-zod/src/generated/types/outfitLook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { ItemHotspot } from "./itemHotspot";
import type { OutfitItem } from "./outfitItem";

export interface OutfitLook {
  id: string;
  title: string;
  vibe: string;
  totalCost: number;
  items: OutfitItem[];
  styleNotes: string;
  modelImageB64?: string;
  hotspots?: ItemHotspot[];
}