# lib/api-zod/src/generated/types/saveOutfitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { OutfitLook } from "./outfitLook";
import type { UserSizes } from "./userSizes";

export interface SaveOutfitRequest {
  prompt: string;
  look: OutfitLook;
  userSizes?: UserSizes;
}