# lib/api-zod/src/generated/types/upsertProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { UpsertProfileRequestGender } from "./upsertProfileRequestGender";
import type { UpsertProfileRequestSkinTone } from "./upsertProfileRequestSkinTone";
import type { UserSizes } from "./userSizes";

export interface UpsertProfileRequest {
  profileId: string;
  name?: string;
  gender?: UpsertProfileRequestGender;
  age?: number;
  skinTone?: UpsertProfileRequestSkinTone;
  location?: string;
  sizes?: UserSizes;
  stylePreferences?: string[];
  avoidKeywords?: string[];
  likedLookIds?: string[];
}