# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */
import type { UserProfileGender } from "./userProfileGender";
import type { UserProfileSkinTone } from "./userProfileSkinTone";
import type { UserSizes } from "./userSizes";

export interface UserProfile {
  /** UUID that identifies the device/user */
  profileId: string;
  name?: string;
  gender?: UserProfileGender;
  age?: number;
  skinTone?: UserProfileSkinTone;
  /** City or region for local brand discovery */
  location?: string;
  sizes?: UserSizes;
  /** Free-form style keywords the user loves */
  stylePreferences?: string[];
  /** Things the user wants to avoid */
  avoidKeywords?: string[];
  likedLookIds?: string[];
  createdAt?: Date;
  updatedAt?: Date;
}