# lib/api-zod/src/generated/types/userSizes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fit Finder API
 * OpenAPI spec version: 0.3.0
 */

export interface UserSizes {
  top?: string;
  bottom?: string;
  shoes?: string;
  dress?: string;
}